refactor: move `diff_roller` main logic out of `pkg/db` since it definitely does not belong there
- It does not really belong where I put it, either, but it's less wrong
